+++ packages/ui/components/overlays/test/OverlayController.test.js
@@ -21,7 +21,7 @@ import { createShadowHost } from '../test-helpers/createShadowHost.js';
  * @typedef {import('../types/OverlayConfig.js').ViewportPlacement} ViewportPlacement
  */
 
-const wrappingDialogNodeStyle = 'display: none; z-index: 9999;';
+const wrappingDialogNodeStyle = 'display: none; z-index: 9999; padding: 0px;';
 
 /**
  * Make sure that all browsers serialize html in a similar way
@@ -45,6 +45,19 @@ function getProtectedMembers(overlayControllerEl) {
     renderTarget,
   };
 }
+
+/**
+ * @param {HTMLElement} element
+ * */
+const isInViewport = element => {
+  const rect = element.getBoundingClientRect();
+  return (
+    rect.top >= 0 &&
+    rect.left >= 0 &&
+    rect.bottom <= (window.innerHeight || document.documentElement.clientHeight) &&
+    rect.right <= (window.innerWidth || document.documentElement.clientWidth)
+  );
+};
 
 const withGlobalTestConfig = () =>
   /** @type {OverlayConfig} */ ({
@@ -905,9 +918,10 @@ describe('OverlayController', () => {
 
         await ctrl.hide();
         expect(document.activeElement).to.equal(input);
+        expect(isInViewport(input)).to.be.true;
       });
 
-      it('supports elementToFocusAfterHide option when shadowRoot involved involved', async () => {
+      it('supports elementToFocusAfterHide option when shadowRoot involved', async () => {
         const input = /** @type {HTMLElement} */ (await fixture('<input />'));
         const contentNode = /** @type {HTMLElement} */ (
           await fixture('<div><textarea></textarea></div>')