+++ packages/ui/components/validate-messages/translations/bg.js
@@ -17,6 +17,8 @@ export default {
       'Моля, въведете {fieldName} между {params.min, date, YYYYMMDD} и {params.max, date, YYYYMMDD}.',
     IsDateDisabled: 'Тази дата не е на разположение, моля, изберете друга.',
     IsEmail: 'Моля, въведете валиден {fieldName} с формат "name@example.com".',
+    MatchesOption:
+      'Не са открити съответстващи резултати. Моля, опитайте с друга ключова дума или категория.',
   },
   warning: {
     Required: 'Моля, въведете също {fieldName}.',
@@ -35,6 +37,8 @@ export default {
       'Моля, въведете {fieldName} между {params.min, date, YYYYMMDD} и {params.max, date, YYYYMMDD}.',
     IsDateDisabled: 'Тази дата не е на разположение, моля, изберете друга.',
     IsEmail: 'Моля, въведете валиден {fieldName} с формат "name@example.com".',
+    MatchesOption:
+      'Не са открити съответстващи резултати. Моля, опитайте с друга ключова дума или категория.',
   },
   success: {
     DefaultOk: 'Добре',