+++ packages/ui/components/validate-messages/translations/nl.js
@@ -17,6 +17,8 @@ export default {
       'Vul een {fieldName} in tussen {params.min, date, YYYYMMDD} en {params.max, date, YYYYMMDD}.',
     IsDateDisabled: 'Deze datum is niet mogelijk, kies een andere.',
     IsEmail: 'Vul een {fieldName} in formaat "name@example.com".',
+    MatchesOption:
+      'Geen overeenkomende resultaten. Probeer een ander trefwoord of een andere categorie.',
   },
   warning: {
     Required: 'Vul een {fieldName} in.',
@@ -35,6 +37,8 @@ export default {
       'Vul een {fieldName} in tussen {params.min, date, YYYYMMDD} en {params.max, date, YYYYMMDD}.',
     IsDateDisabled: 'Deze datum is niet mogelijk, kies een andere.',
     IsEmail: 'Vul een {fieldName} in formaat "name@example.com".',
+    MatchesOption:
+      'Geen overeenkomende resultaten. Probeer een ander trefwoord of een andere categorie.',
   },
   success: {
     DefaultOk: 'Okee',